+++ src/test/php/d4np/utils/Container/Fixture/AbstractThing.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace D4np\Utils\Tests\Container\Fixture;
+
+abstract class AbstractThing
+{
+}